+++ src/Record/Record.Model/Utils/TimeUtils.cs
@@ -1,0 +1,24 @@
+﻿using VDS.RDF;
+
+namespace Records.Utils;
+
+public static class TimeUtils
+{
+    public static List<Triple?> CreateHasTimeTriples(UriNode timeHaver, DateTime dateTime)
+    {
+        var blankNode = new BlankNode(dateTime.ToString());
+
+        var gYear = Namespaces.Time.UriNodes.GetYearLiteralNode(dateTime);
+        var gregMonth = Namespaces.Greg.UriNodes.GetGregorianMonthUriNode(dateTime);
+        var gDay = Namespaces.Time.UriNodes.GetDayLiteralNode(dateTime);
+
+        return
+        [
+            new(timeHaver, Namespaces.Time.UriNodes.HasTime, blankNode),
+            new(blankNode, Namespaces.Rdf.UriNodes.Type, Namespaces.Time.UriNodes.DateTimeDescription),
+            new(blankNode, Namespaces.Time.UriNodes.Year, gYear),
+            new(blankNode, Namespaces.Time.UriNodes.Month, gregMonth),
+            new(blankNode, Namespaces.Time.UriNodes.Day, gDay)
+        ];
+    }
+}
